Remove all references to Parse
Since we don't really integrate with them anyway.
This is related to
https://github.com/e-mission/e-mission-data-collection/pull/184/commits/ebe2d20e6e12d56a7045d4d2e93c61ab6cca89bf
in PR
https://github.com/e-mission/e-mission-data-collection/pull/184

Note that there appears to be a subtle bug here in which we save the sync
config but don't update the profile.  And we use the profile to determine the
frequency at which to send the IOS pushes. But that is not really related to
this change which attempted to save channels to a non-existent server.